Add files via upload
JSON file is a subset of JSON file from Moral Foundations Twitter Corpus: https://osf.io/k5n7y/#!

Due to privacy reasons, the JSON files don't contain text or user_id of the authors of the tweets. You would need access to twitter API, with credentials to extract tweets from it, with the help of all the tweet_ids mentioned in it.